requirements.txt
einops==0.6.0
MedPy==0.4.0
opencv-python==4.6.0.66
pandas==1.5.2
scipy==1.9.1
SimpleITK==2.2.1
torchvision==0.13.1

numpy==1.23.3
matplotlib==3.6.2
nibabel==4.0.2
torchio==0.18.90
scikit-image==0.19.2
scikit-learn==1.2.0
imageio==2.22.0
tqdm==4.64.1
tensorboardx==2.5.1
pyyaml==6.0
monai==1.2.0
setuptools==63.4.1
seaborn==0.11.2
statannotations==0.5.0
timm==0.3.2